animated background in own component
