+++ src/app/waitlist/WaitlistPage.tsx
@@ -1,14 +1,15 @@
+"use client";
+
 import { useState } from "react";
-import { Button } from "./ui/button";
-import { Input } from "./ui/input";
-import Logo from "./ui/Logo";
+import { useRouter } from "next/navigation";
 
-interface WaitlistPageProps {
-  onBack: () => void;
-  onSubmit: () => void;
-}
+import { Button } from "../../components/ui/button";
+import { Input } from "../../components/ui/input";
+import Logo from "../..//components/ui/Logo";
+import { AnimatedBackground } from "../../components/AnimatedBackground";
 
-export function WaitlistPage({ onBack, onSubmit }: WaitlistPageProps) {
+export function WaitlistPage() {
+  const router = useRouter();
   const [email, setEmail] = useState("");
   const [loading, setLoading] = useState(false);
 
@@ -24,15 +25,14 @@ export function WaitlistPage({ onBack, onSubmit }: WaitlistPageProps) {
       });
 
       const data = await res.json();
-
       if (!res.ok) throw new Error(data.error || "Something went wrong");
 
-      onSubmit(); // navigation happens here
+      router.push("/waitlist/confirmation");
     } catch (err) {
       console.error("Email send failed:", err);
       alert("Oops, something went wrong. Please try again.");
     } finally {
-      setLoading(false); // reset loading state
+      setLoading(false);
     }
   };
 
@@ -45,7 +45,7 @@ export function WaitlistPage({ onBack, onSubmit }: WaitlistPageProps) {
           <button
             type="button"
             aria-label="Go back"
-            onClick={onBack}
+            onClick={() => router.push("/")}
             className="h-12 w-auto"
           >
             <div className="h-full aspect-[1000/600]">
@@ -119,75 +119,7 @@ export function WaitlistPage({ onBack, onSubmit }: WaitlistPageProps) {
         </div>
       </div>
 
-      {/* Right side - Dark background */}
-      <div className="flex-1 bg-gray-900 relative overflow-hidden hidden lg:block">
-        {/* Enhanced visual texture with animation */}
-        <div className="absolute inset-0">
-          <div className="absolute inset-0 bg-gradient-to-br from-gray-800 via-gray-900 to-black"></div>
-          <div
-            className="absolute inset-0"
-            style={{
-              background:
-                "radial-gradient(circle at 30% 40%, rgba(255,255,255,0.15) 0%, transparent 30%)",
-              animation: "float1 9s ease-in-out infinite",
-            }}
-          ></div>
-          <div
-            className="absolute inset-0"
-            style={{
-              background:
-                "radial-gradient(circle at 70% 60%, rgba(255,255,255,0.08) 0%, transparent 65%)",
-              animation: "float2 11s ease-in-out infinite",
-            }}
-          ></div>
-          <div
-            className="absolute inset-0"
-            style={{
-              background:
-                "radial-gradient(circle at 50% 50%, rgba(255,255,255,0.04) 0%, transparent 90%)",
-              animation: "float3 13s ease-in-out infinite",
-            }}
-          ></div>
-        </div>
-
-        {/* Subtle grid pattern */}
-        <div className="absolute inset-0 opacity-[0.03]">
-          <div
-            className="absolute inset-0"
-            style={{
-              backgroundImage: `
-              linear-gradient(rgba(255,255,255,0.1) 1px, transparent 1px),
-              linear-gradient(90deg, rgba(255,255,255,0.1) 1px, transparent 1px)
-            `,
-              backgroundSize: "50px 50px",
-            }}
-          ></div>
-        </div>
-
-        {/* Animation keyframes */}
-        <style>{`
-          @keyframes float1 {
-            0%, 100% { transform: translate(0, 0) scale(1.3); }
-            25% { transform: translate(25px, -30px) scale(1.5); }
-            50% { transform: translate(-15px, 25px) scale(2); }
-            75% { transform: translate(-25px, -10px) scale(1.35); }
-          }
-          
-          @keyframes float2 {
-            0%, 100% { transform: translate(0, 0) scale(1.3); }
-            33% { transform: translate(-30px, 20px) scale(1.55); }
-            66% { transform: translate(35px, -25px) scale(1); }
-          }
-          
-          @keyframes float3 {
-            0%, 100% { transform: translate(0, 0) scale(1.3); }
-            20% { transform: translate(15px, 35px) scale(1.4); }
-            40% { transform: translate(-25px, -20px) scale(1); }
-            60% { transform: translate(30px, 10px) scale(1.5); }
-            80% { transform: translate(-10px, -30px) scale(1); }
-          }
-        `}</style>
-      </div>
+      <AnimatedBackground />
     </div>
   );
 }
