+++ src/app/waitlist/confirmation/page.tsx
@@ -1,0 +1,10 @@
+"use client";
+
+import { useRouter } from "next/navigation";
+import { ConfirmationPage } from "./ConfirmationPage";
+
+export default function ConfirmationRoute() {
+  const router = useRouter();
+
+  return <ConfirmationPage onHome={() => router.push("/")} />;
+}
